+++ R/class_file.R
@@ -143,7 +143,13 @@ file_info <- function(files) {
 }
 
 file_time <- function(info) {
-  digest_obj64(max(c(-Inf, replace_na(as.numeric(info$mtime), -Inf))))
+  diff <- difftime(
+    time1 = info$mtime,
+    time2 = file_time_reference,
+    units = "days",
+    tz = "UTC"
+  )
+  file_diff_chr(max(replace_na(c(as.numeric(diff), 0), 0)))
 }
 
 file_bytes <- function(info) {
@@ -153,4 +159,27 @@ file_bytes <- function(info) {
 
 file_size <- function(bytes) {
   digest_obj64(bytes)
+}
+
+file_diff_chr <- function(dbl) {
+  sprintf("t%ss", as.character(dbl))
+}
+
+file_diff_dbl <- function(chr) {
+  as.numeric(gsub(pattern = "^t|s$", replacement = "", x = chr))
+}
+
+file_time_posixct <- function(chr) {
+  diff <- as.difftime(file_diff_dbl(chr), units = "days")
+  file_time_system_tz(diff + file_time_reference)
+}
+
+file_time_reference <- as.POSIXct(
+  "1970-01-01 00:00:00",
+  format = "%Y-%m-%d %H:%M:%S",
+  tz = "UTC"
+)
+
+file_time_system_tz <- function(x) {
+  as.POSIXct(as.POSIXlt(x, tz = Sys.timezone()))
 }