+++ NEWS.md
@@ -32,6 +32,7 @@
 * Make the reporters show when a pattern ends if the pattern built at least one target and none of the targets errored or canceled.
 * Use words "start" and "built" in reporters.
 * Use the region of the AWS S3 bucket instead of the local `AWS_DEFAULT_REGION` environment variable (`check_region = TRUE`; #400, @tomsing1).
+* In `tar_meta()`, return `POSIXct` times in the time zone of the calling system (#131).
 
 # targets 0.3.1